reformating

+++ src/Searcher.h
@@ -38,12 +38,12 @@ namespace Searcher {
     public:
 
         GameClock gameclock[CLR_NO];
-        std::vector<Move> root_moves;   // search these moves only restrict
+        std::vector<Move> root_moves;   // restrict search to these moves only
 
         u32  movetime;  // search <x> time in milli-seconds
         u08  movestogo; // search <x> moves to the next time control
         u08  depth;     // search <x> depth (plies) only
-        u32  nodes;     // search <x> nodes only
+        u64  nodes;     // search <x> nodes only
         u08  mate;      // search mate in <x> moves
         bool infinite;  // search until the "stop" command
         bool ponder;    // search on ponder move
